src/third_party/blink/web_tests/fast/shapedetection/shapedetection-creation.html

<!DOCTYPE html>
<script src=../../resources/testharness.js></script>
<script src=../../resources/testharnessreport.js></script>
<script>

// This test verifies that DetectedText can be created
test(function() {
  var detectedText = new DetectedText();
  assert_true(detectedText instanceof DetectedText);
}, 'DetectedText instance can be created.');

// This test verifies that TextDetector can be created
test(function() {
  var textDetector = new TextDetector();
  assert_true(textDetector instanceof TextDetector);
}, 'TextDetector instance can be created.');

</script>
